+++ packages/plugin-instructions/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jspsych/plugin-instructions
 
+## 1.1.3
+
+### Patch Changes
+
+- [#3039](https://github.com/jspsych/jsPsych/pull/3039) [`347bbb55`](https://github.com/jspsych/jsPsych/commit/347bbb557c4cf954e471e8d66188db0b28e364be) Thanks [@jodeleeuw](https://github.com/jodeleeuw)! - Fix simulation mode behavior so that setting RT and/or view_history correctly sets the other parameter
+
 ## 1.1.2
 
 ### Patch Changes